requirements.txt
prometheus_client
python-socketio[client]
python-dotenv
requests
